+++ src/main/java/insane96mcp/survivalreimagined/module/mining/feature/MiningMisc.java
@@ -5,15 +5,24 @@ import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
+import insane96mcp.insanelib.util.MCUtils;
 import insane96mcp.survivalreimagined.module.Modules;
+import net.minecraft.world.entity.ai.attributes.AttributeModifier;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.block.AbstractSkullBlock;
 import net.minecraft.world.level.block.InfestedBlock;
+import net.minecraftforge.common.ForgeMod;
+import net.minecraftforge.event.entity.EntityJoinLevelEvent;
 import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
+
+import java.util.UUID;
 
 @Label(name = "Misc", description = "Various mining changes")
 @LoadFeature(module = Modules.Ids.MINING)
 public class MiningMisc extends Feature {
+
+	public static final UUID BLOCK_REACH_REDUCTION_UUID = UUID.fromString("bae34f6a-c58e-4622-b2ab-f1b89b73b781");
 
 	@Config
 	@Label(name = "Insta-Mine Silverfish", description = "Silverfish blocks will insta-mine like pre-1.17")
@@ -23,8 +32,12 @@ public class MiningMisc extends Feature {
 	@Label(name = "Insta-Mine Heads", description = "Heads will insta-break")
 	public static Boolean instaMineHeads = true;
 
+	@Config(min = -4, max = 4)
+	@Label(name = "Mining Range reduction", description = "Reduce the range at which players can mine")
+	public static Double miningRangeReduction = -1d;
+
 	//TODO Prevent swords from mining 0 hardness blocks
-	/*@Config
+    /*@Config
 	@Label(name = "No Sword breaking insta-mine blocks", description = "Prevents swords from breaking blocks like grass, etc.")
 	public static Boolean noSwordBreaking = true;*/
 
@@ -37,7 +50,7 @@ public class MiningMisc extends Feature {
 		if (!this.isEnabled())
 			return;
 
-		/*if (event.getEntity().getMainHandItem().getItem() instanceof SwordItem && event.getState().destroySpeed == 0f)
+        /*if (event.getEntity().getMainHandItem().getItem() instanceof SwordItem && event.getState().destroySpeed == 0f)
 			event.setNewSpeed(0f);*/
 
 		silverfishBreakSpeed(event);
@@ -58,5 +71,15 @@ public class MiningMisc extends Feature {
 
 		if (event.getState().getBlock() instanceof AbstractSkullBlock)
 			event.setNewSpeed(Float.MAX_VALUE);
+	}
+
+	@SubscribeEvent
+	public void onPlayerJoinLevel(EntityJoinLevelEvent event) {
+		if (!this.isEnabled()
+				|| miningRangeReduction == 0d
+				|| !(event.getEntity() instanceof Player player))
+			return;
+
+		MCUtils.applyModifier(player, ForgeMod.BLOCK_REACH.get(), BLOCK_REACH_REDUCTION_UUID, "Block reach reduction", miningRangeReduction, AttributeModifier.Operation.ADDITION, false);
 	}
 }